refactor: rename folders containing dtos to dtos
